FE-Feat: 네이밍 변경(Badge)
