+++ packages/ui/index.tsx
@@ -2,6 +2,6 @@ export { default as Toast } from "@ui/src/components/common/Toast/index";
 export { default as notify } from "@ui/src/components/common/Toast/notify";
 export { default as Modal } from "@ui/src/components/common/Modal/index";
 
-export { default as Label } from "@ui/src/components/common/Label/index";
+export { default as Badge } from "@ui/src/components/common/Badge/index";
 
 // export * from "@ui/components/ui/button"
